% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bigcat.R
\name{bigcat}
\alias{bigcat}
\title{Print a string surrounded by stars}
\usage{
bigcat(x)
}
\arguments{
\item{x}{A character vector with one element.}
}
\description{
Print a string surrounded by stars
}
\examples{
bigcat("Load the data")
}
