% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot.R
\name{ggiplot}
\alias{ggiplot}
\title{ggiplot with bergr theme and Dark2 color scheme}
\usage{
ggiplot(...)
}
\arguments{
\item{...}{Arguments to be passed on to ggiplot::ggiplot.}
}
\description{
ggiplot with bergr theme and Dark2 color scheme
}
