% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot.R
\name{ggplot}
\alias{ggplot}
\title{ggplot with bergr theme and Dark2 color scheme}
\usage{
ggplot(...)
}
\arguments{
\item{...}{Arguments to be passed on to ggplot2::ggplot.}
}
\description{
ggplot with bergr theme and Dark2 color scheme
}
