% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palette_bergr.R
\docType{data}
\name{palette_bergr}
\alias{palette_bergr}
\title{Custom palette with small modification to Dark2.}
\format{
An object of class \code{character} of length 8.
}
\usage{
palette_bergr
}
\description{
Custom palette with small modification to Dark2.
}
\keyword{datasets}
