% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_plots.R
\name{save_plots}
\alias{save_plots}
\title{Recursively save elements of a list of ggplots}
\usage{
save_plots(x, root, ext = "png", verbose = T, remove_title = F, ...)
}
\arguments{
\item{x}{A list of ggplots that you wish to save to disk as individual files.}

\item{root}{Character of length 1 that specifies the root directory. Your files will be saved in \code{root} or in subdirectories thereof.}

\item{ext}{Character of length 1 that specifies the file extension. "png" by default.}

\item{verbose}{Set to FALSE if you do not wish to print file paths as they save.}

\item{remove_title}{Set to TRUE if you wish to remove titles from each plot.}

\item{...}{Arguments to pass to ggplot.}
}
\description{
Save a nested list of ggplots in a directory with the same structure.
}
\examples{
# Make list of plots
library(ggplot2)
library(fs)
p1 <- ggplot(head(mtcars), aes(x = cyl, y = mpg))
p2 <- ggplot(tail(mtcars), aes(x = cyl, y = mpg))
my_list <- list(
  a = p1,
  b = list(c = p2)
)
# Set temporary root directory
temp <- tempdir()
root <- paste0(temp, "/plots")
dir.create(root, showWarnings = FALSE)
# Save files
save_plots(my_list, root = root, height = 4, width = 6)
dir_tree(root)
}
