% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_tabs.R
\name{save_tabs}
\alias{save_tabs}
\title{Recursively save elements of a list of strings}
\usage{
save_tabs(x, root, ext = "tex", verbose = T)
}
\arguments{
\item{x}{A list of character vectors that you wish to save to disk as individual files.}

\item{root}{Character of length 1 that specifies the root directory. Your files will be saved in \code{root} or in subdirectories thereof.}

\item{ext}{Character of length 1 that specifies the file extension. "tex" by default.}

\item{verbose}{Set to FALSE if you do not wish to print file paths as they save.}
}
\description{
Save a nested list of strings in a directory with the same structure. Works great for latex tables!
}
\examples{
# Make list of tables
library(bergr)
library(knitr)
library(fs)
d1 <- head(mtcars)
d2 <- tail(mtcars)
my_list <- list(
  a = knitr::kable(d1, format = "latex"),
  b = list(c = knitr::kable(d2, format = "latex"))
)
# Set temporary root directory
temp <- tempdir()
root <- paste0(temp, "/tabs")
dir.create(root, showWarnings = FALSE)
# Save files
save_tabs(my_list, root = root)
dir_tree(root)
}
