% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coefplot.R
\name{coefplot_wide}
\alias{coefplot_wide}
\title{Custom coefficient plot for many fixest models}
\usage{
coefplot_wide(models, treat_var, var_labels = NULL)
}
\arguments{
\item{models}{one or more fixest models}

\item{treat_var}{a single independent variable.
\code{coefplot_wide()} returns the coefficient on this variable}

\item{var_labels}{a named vector of dependent variables, one for each model.
The dependent variable is labeled according to its label.}
}
\description{
Custom coefficient plot for many fixest models
}
